Handle printing of child objects for JSON pretty-print
Previously, any child objects under the object provided to the
objectToJson function for pretty-printing were omitted. With the changes
in this commit, the list of alphabetized keys supplied to JSON.stringify
for pretty printing includes keys found in the child objects. This
allows for the entire object structure to be pretty-printed, with keys
in the output being sorted alphabetically.