Update dxl-client dependency and Travis provisioning config
This commit bumps the minimum @opendxl/dxl-client dependency to 0.1.0
and updates the Travis CI configuration to use the client's
provisionconfig CLI command to provision the client configuration used
when running tests. This replaces the prior Travis configuration which
relied upon a Docker volume mount to use the broker's own keystore as
the client configuration for tests.